Move UI test utils to Service test utils
